Add additional vars to used_cols list
For some queries, such as UPDATE, the vars in the query tree are not
enough. We need all of the columns in the table for the current
implementation of UPDATE when the table being updated has a constraint
(amongst other possible scenarios). So, we can get the additional
columns from a processed (pull_up_var_clause) version of
root->processed_tlist.

Note that this is an experiment which worked but we're not sure if this
is a good idea/all the ramifications and how everything got here

Also, interpret varattno as don't bother with proj list