Populate childrel's used cols
For partition tables, UNION ALL subqueries, and some other cases, the
used columns are available in the AppendRelInfos and not in the
targetlist of the main query tree. Unfortunately, the AppendRelInfos
seem to always include all columns present in child relations. So, if we
want to only get used columns, we need to use a top-level information
structure -- in this case we mostly use simple_rel_array and
simple_rte_array to know which of the columns are needed for the query.
So, we use the AppendRelInfo only to know if the relation is inherited
or not. This still does not handle some cases -- such as
table type.