Get the child partition rtables
Ashwin and I spent some time figuring out why getting the columns for an UPDATE
to a partition table during planning was not working. It turns out that
inheritance_planner makes a new PlannerInfo for all child partitions and then
makes a path for each one and then adds these to a list which becomes a pathlist
in ModifyTable node. While doing planning for each child partition, the
used_cols would only be populated for that relation's RTE. Though this list of
PlannerInfos containing all of the separate RTEs with the used_cols filled in is
passed to the ModifyTable node, the final rangetable for the whole query (the
original, top-level UPDATE query) is actually just copied from the
RangeTable of the first child partition. It seems like other components of the
PlannerInfos get copied over to the main query tree as needed, but, the modified
RangeTableEntries themselves are not. We were wondering how it can be that no
other part of planning needs the rtables from each of the relation's query
trees, but, it seems like some kind of temporary thing only used for this phase
in planning. We need to do more investigation to understand if we can either get
the used_cols from the rangetables of the children in the ModifyTable node (if
those are available later in planning) or if we should make some change to
inheritance_planner to actually copy over all of the information from the child
rtables to the top level one (or if we should just wait until David Rowley
changes all of this code completely anyway).