Don't use appendrelinfo translated_vars and copy list
It turns out that simple_rel_array's RelOptInfos' PathTargets have most
everything needed and it is not necessary to go through AppendRelInfos'
translated_vars. Also, pull_vars_of_level on the parse tree is also not
needed.

Separately, we found we needed to copy the list because in
inheritance_planner all of the PlannerInfos we create start out as a
copy of the main PlannerInfo and then we modify them. The
RangeTableEntries' used_cols lists get modified differently for
different copies of the PlannerInfo, so, we can't concatenate one to
another since they are all basically pieces of the same list, so, make a
copy.
Will replace later with bitmap.